requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pg8000>=1.31
pydantic[email]>=2.0
bcrypt>=4.0
PyJWT>=2.8
